// src/types/generated/ProviderDefinition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProviderModel } from "./ProviderModel";
import type { ProviderProtocol } from "./ProviderProtocol";

export type ProviderDefinition = { id: string, name: string, api: string, npm: string, doc?: string, env: Array<string>, protocol: ProviderProtocol, models: Array<ProviderModel>, defaultModel?: string, defaultApi?: string, isApiOverridden: boolean, isCustom: boolean, isLocal: boolean, };